Actividad-1-Desarrollo-del-Proceso-ETL/etl/requirements.txt
# ETL System - Dependencias
# ─────────────────────────
pydantic>=2.0
python-dotenv>=1.0
aiohttp>=3.9
pyodbc>=5.0
streamlit>=1.30
pandas>=2.0
plotly>=5.18

